Move painting squares function to a component
